Restructure dir and add rollup
