--- quiz-us-components/Card/README.md
@@ -1,19 +1,0 @@
-# Card
-
-## Description
-
-A card is used during flash card studies.
-
-## Usage
-
-```js
-import Card from '@quiz-us/components/Card';
-
-() => <Card isFlipped={false} />;
-```
-
-## Properties
-
-| Props       | Type      | Required | Values | Default | Description                        |
-| ----------- | --------- | :------: | ------ | ------- | ---------------------------------- |
-| `isFlipped` | `boolean` |    -     |        | `false` | whether or not the card is flipped |
