+++ quiz-us-components/package.json
@@ -1,10 +1,11 @@
 {
   "name": "@quiz-us/components",
-  "version": "1.0.0",
+  "version": "1.0.2",
   "description": "Component library for formative assessments",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "npx rollup -c"
   },
   "keywords": [],
   "author": "",
@@ -13,7 +14,7 @@
     "@material-ui/core": "^3.9.3",
     "@material-ui/styles": "^3.0.0-alpha.10",
     "react": "^16.8.6",
-    "prop-types": "^15.7.2"
-  },
-  "dependencies": {}
+    "prop-types": "^15.7.2",
+    "react-dom": "^16.8.6"
+  }
 }
