+++ quiz-us-components/src/index.js
@@ -1,2 +1,1 @@
-export { default as Card } from './Card';
 export { default as QuestionMultipleChoice } from './QuestionMultipleChoice';
