--- stories/Card/Card.js
@@ -1,4 +1,0 @@
-import React from 'react';
-import Card from '../../quiz-us-components/Card';
-
-export default () => <Card isFlipped={false} />;
