+++ stories/index.stories.js
@@ -1,5 +1,3 @@
-import Card from './Card';
 import Questions from './Questions';
 
-Card(module);
 Questions(module);
